requirements.txt
connexion
connexion[swagger-ui]
Flask-Caching
flask_marshmallow
flask_sqlalchemy
Flask-SQLAlchemy-Caching
folium
marshmallow-sqlalchemy
newrelic
pandas
redis
waitress
